# api/v1alpha1/engine/types.gen.go
// Package engine provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package engine

// Defines values for EvaluateResponseStatus.
const (
	APPROVED EvaluateResponseStatus = "APPROVED"
	MODIFIED EvaluateResponseStatus = "MODIFIED"
)

// Error defines model for Error.
type Error struct {
	// Detail Detailed error message
	Detail *string `json:"detail,omitempty"`

	// Instance Request identifier
	Instance *string `json:"instance,omitempty"`

	// Status HTTP status code
	Status int32 `json:"status"`

	// Title Short error summary
	Title string `json:"title"`

	// Type Error type identifier
	Type string `json:"type"`
}

// EvaluateRequest defines model for EvaluateRequest.
type EvaluateRequest struct {
	ServiceInstance ServiceInstance `json:"service_instance"`
}

// EvaluateResponse defines model for EvaluateResponse.
type EvaluateResponse struct {
	EvaluatedServiceInstance ServiceInstance `json:"evaluated_service_instance"`

	// SelectedProvider Service provider selected by policies
	SelectedProvider string `json:"selected_provider"`

	// Status APPROVED - Request unchanged by policies
	// MODIFIED - Request was modified by policies
	Status EvaluateResponseStatus `json:"status"`
}

// EvaluateResponseStatus APPROVED - Request unchanged by policies
// MODIFIED - Request was modified by policies
type EvaluateResponseStatus string

// ServiceInstance defines model for ServiceInstance.
type ServiceInstance struct {
	// Spec Service specification (flexible schema)
	Spec map[string]interface{} `json:"spec"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Forbidden defines model for Forbidden.
type Forbidden = Error

// InternalServerError defines model for InternalServerError.
type InternalServerError = Error

// PolicyConflict defines model for PolicyConflict.
type PolicyConflict = Error

// Rejected defines model for Rejected.
type Rejected = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// EvaluateRequestJSONRequestBody defines body for EvaluateRequest for application/json ContentType.
type EvaluateRequestJSONRequestBody = EvaluateRequest