# internal/api/engine/server.gen.go
// Package engineserver provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package engineserver

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for EvaluateResponseStatus.
const (
	APPROVED EvaluateResponseStatus = "APPROVED"
	MODIFIED EvaluateResponseStatus = "MODIFIED"
)

// Error defines model for Error.
type Error struct {
	// Detail Detailed error message
	Detail *string `json:"detail,omitempty"`

	// Instance Request identifier
	Instance *string `json:"instance,omitempty"`

	// Status HTTP status code
	Status int32 `json:"status"`

	// Title Short error summary
	Title string `json:"title"`

	// Type Error type identifier
	Type string `json:"type"`
}

// EvaluateRequest defines model for EvaluateRequest.
type EvaluateRequest struct {
	ServiceInstance ServiceInstance `json:"service_instance"`
}

// EvaluateResponse defines model for EvaluateResponse.
type EvaluateResponse struct {
	EvaluatedServiceInstance ServiceInstance `json:"evaluated_service_instance"`

	// SelectedProvider Service provider selected by policies
	SelectedProvider string `json:"selected_provider"`

	// Status APPROVED - Request unchanged by policies
	// MODIFIED - Request was modified by policies
	Status EvaluateResponseStatus `json:"status"`
}

// EvaluateResponseStatus APPROVED - Request unchanged by policies
// MODIFIED - Request was modified by policies
type EvaluateResponseStatus string

// ServiceInstance defines model for ServiceInstance.
type ServiceInstance struct {
	// Spec Service specification (flexible schema)
	Spec map[string]interface{} `json:"spec"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Forbidden defines model for Forbidden.
type Forbidden = Error

// InternalServerError defines model for InternalServerError.
type InternalServerError = Error

// PolicyConflict defines model for PolicyConflict.
type PolicyConflict = Error

// Rejected defines model for Rejected.
type Rejected = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// EvaluateRequestJSONRequestBody defines body for EvaluateRequest for application/json ContentType.
type EvaluateRequestJSONRequestBody = EvaluateRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Evaluate request payload against policies
	// (POST /policies:evaluateRequest)
	EvaluateRequest(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Evaluate request payload against policies
// (POST /policies:evaluateRequest)
func (_ Unimplemented) EvaluateRequest(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// EvaluateRequest operation middleware
func (siw *ServerInterfaceWrapper) EvaluateRequest(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.EvaluateRequest(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/policies:evaluateRequest", wrapper.EvaluateRequest)
	})

	return r
}

type BadRequestJSONResponse Error

type ForbiddenJSONResponse Error

type InternalServerErrorJSONResponse Error

type PolicyConflictJSONResponse Error

type RejectedJSONResponse Error

type UnauthorizedJSONResponse Error

type EvaluateRequestRequestObject struct {
	Body *EvaluateRequestJSONRequestBody
}

type EvaluateRequestResponseObject interface {
	VisitEvaluateRequestResponse(w http.ResponseWriter) error
}

type EvaluateRequest200JSONResponse EvaluateResponse

func (response EvaluateRequest200JSONResponse) VisitEvaluateRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type EvaluateRequest400JSONResponse struct{ BadRequestJSONResponse }

func (response EvaluateRequest400JSONResponse) VisitEvaluateRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type EvaluateRequest401JSONResponse struct{ UnauthorizedJSONResponse }

func (response EvaluateRequest401JSONResponse) VisitEvaluateRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type EvaluateRequest403JSONResponse struct{ ForbiddenJSONResponse }

func (response EvaluateRequest403JSONResponse) VisitEvaluateRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)

	return json.NewEncoder(w).Encode(response)
}

type EvaluateRequest406JSONResponse struct{ RejectedJSONResponse }

func (response EvaluateRequest406JSONResponse) VisitEvaluateRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(406)

	return json.NewEncoder(w).Encode(response)
}

type EvaluateRequest409JSONResponse struct{ PolicyConflictJSONResponse }

func (response EvaluateRequest409JSONResponse) VisitEvaluateRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type EvaluateRequest500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response EvaluateRequest500JSONResponse) VisitEvaluateRequestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Evaluate request payload against policies
	// (POST /policies:evaluateRequest)
	EvaluateRequest(ctx context.Context, request EvaluateRequestRequestObject) (EvaluateRequestResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// EvaluateRequest operation middleware
func (sh *strictHandler) EvaluateRequest(w http.ResponseWriter, r *http.Request) {
	var request EvaluateRequestRequestObject

	var body EvaluateRequestJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.EvaluateRequest(ctx, request.(EvaluateRequestRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "EvaluateRequest")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(EvaluateRequestResponseObject); ok {
		if err := validResponse.VisitEvaluateRequestResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}